Revisions based on initial feedback
