+++ client/blocks/jetpack-benefits/benefit-card.tsx
@@ -4,6 +4,11 @@
 import React from 'react';
 import { TranslateResult } from 'i18n-calypso';
 
+/**
+ * Internal Dependencies
+ */
+import classNames from 'classnames';
+
 interface Props {
 	stat?: TranslateResult | string | number | null;
 	headline: TranslateResult;
@@ -23,17 +28,15 @@ export const JetpackBenefitsCard: React.FC< Props > = ( {
 			return <span className="jetpack-benefits__card-stat">{ stat }</span>;
 		}
 	};
+	const statBlockClassNames = classNames( 'jetpack-benefits__stat-block ', {
+		'jetpack-benefits__stat-block--placeholder': placeholder,
+	} );
 
 	return (
 		<div className="jetpack-benefits__card card">
 			<div className="jetpack-benefits__card-content">
 				<p className="jetpack-benefits__card-headline">{ headline }</p>
-				<div
-					className={
-						'jetpack-benefits__stat-block ' +
-						( placeholder ? 'jetpack-benefits__stat-block--placeholder' : '' )
-					}
-				>
+				<div className={ statBlockClassNames }>
 					{ getStat() }
 					<div className="jetpack-benefits__card-description">{ description }</div>
 				</div>
