+++ client/blocks/jetpack-benefits/index.tsx
@@ -11,21 +11,7 @@ import { useTranslate } from 'i18n-calypso';
 import {
 	isJetpackBackupSlug,
 	isJetpackScanSlug,
-	isJetpackAntiSpamSlug,
 	isJetpackPlanSlug,
-	planHasFeature,
-	FEATURE_JETPACK_BACKUP_REALTIME,
-	FEATURE_JETPACK_BACKUP_DAILY,
-	FEATURE_JETPACK_BACKUP_REALTIME_MONTHLY,
-	FEATURE_JETPACK_BACKUP_DAILY_MONTHLY,
-	FEATURE_JETPACK_SCAN_DAILY,
-	FEATURE_JETPACK_SCAN_DAILY_MONTHLY,
-	FEATURE_JETPACK_SEARCH,
-	FEATURE_JETPACK_SEARCH_MONTHLY,
-	FEATURE_JETPACK_ANTI_SPAM,
-	FEATURE_JETPACK_ANTI_SPAM_MONTHLY,
-	JETPACK_SEARCH_PRODUCTS,
-	FEATURE_ALL_PREMIUM_FEATURES_JETPACK,
 } from '@automattic/calypso-products';
 import getRewindState from 'calypso/state/selectors/get-rewind-state';
 import getSiteScanState from 'calypso/state/selectors/get-site-scan-state';
@@ -33,66 +19,18 @@ import JetpackBenefitsSiteVisits from 'calypso/blocks/jetpack-benefits/site-visi
 import JetpackBenefitsScanHistory from 'calypso/blocks/jetpack-benefits/scan-history';
 import JetpackBenefitsSiteBackups from 'calypso/blocks/jetpack-benefits/site-backups';
 import { JetpackBenefitsCard } from 'calypso/blocks/jetpack-benefits/benefit-card';
+import {
+	productHasSearch,
+	productHasBackups,
+	productHasScan,
+	productHasActivityLog,
+	productHasAntiSpam,
+} from 'calypso/blocks/jetpack-benefits/feature-checks';
 
 /**
  * Style dependencies
  */
 import './style.scss';
-
-export const productHasBackups = ( productSlug: string ) => {
-	return (
-		// standalone backup product
-		isJetpackBackupSlug( productSlug ) ||
-		// check plans for Jetpack backup features
-		( isJetpackPlanSlug( productSlug ) &&
-			( planHasFeature( productSlug, FEATURE_JETPACK_BACKUP_DAILY ) ||
-				planHasFeature( productSlug, FEATURE_JETPACK_BACKUP_DAILY_MONTHLY ) ||
-				planHasFeature( productSlug, FEATURE_JETPACK_BACKUP_REALTIME ) ||
-				planHasFeature( productSlug, FEATURE_JETPACK_BACKUP_REALTIME_MONTHLY ) ) )
-	);
-};
-
-export const productHasScan = ( productSlug: string ) => {
-	return (
-		// standalone scan product
-		isJetpackScanSlug( productSlug ) ||
-		// check plans for Jetpack scan features
-		( isJetpackPlanSlug( productSlug ) &&
-			( planHasFeature( productSlug, FEATURE_JETPACK_SCAN_DAILY ) ||
-				planHasFeature( productSlug, FEATURE_JETPACK_SCAN_DAILY_MONTHLY ) ) )
-	);
-};
-
-export const productHasSearch = ( productSlug: string ) => {
-	return (
-		// standalone search product
-		// there is not currently a isJetpackSearchSlug
-		JETPACK_SEARCH_PRODUCTS.includes( productSlug ) ||
-		// check plans for Jetpack search features
-		( isJetpackPlanSlug( productSlug ) &&
-			( planHasFeature( productSlug, FEATURE_JETPACK_SEARCH ) ||
-				planHasFeature( productSlug, FEATURE_JETPACK_SEARCH_MONTHLY ) ||
-				// This is a bit obscure - checks specifically for Jetpack Business (Professional)
-				// Is it an error that the plan spec in plans-list.js does not contain search features?
-				planHasFeature( productSlug, FEATURE_ALL_PREMIUM_FEATURES_JETPACK ) ) )
-	);
-};
-
-export const productHasAntiSpam = ( productSlug: string ) => {
-	// check that this product is standalone anti-spam or one of the plans that contains it
-	return (
-		// standalone anti-spam product
-		isJetpackAntiSpamSlug( productSlug ) ||
-		// check plans for anti-spam features
-		( isJetpackPlanSlug( productSlug ) &&
-			( planHasFeature( productSlug, FEATURE_JETPACK_ANTI_SPAM ) ||
-				planHasFeature( productSlug, FEATURE_JETPACK_ANTI_SPAM_MONTHLY ) ) )
-	);
-};
-
-export const productHasActivityLog = ( productSlug: string ) => {
-	return isJetpackPlanSlug( productSlug ) || isJetpackBackupSlug( productSlug );
-};
 
 interface Props {
 	siteId: number;
@@ -118,39 +56,6 @@ const JetpackBenefits: React.FC< Props > = ( { siteId, productSlug } ) => {
 		return 'unavailable' !== scanState?.state;
 	};
 
-	const renderSiteSearch = () => {
-		return (
-			<JetpackBenefitsCard
-				headline={ translate( 'Search' ) }
-				description={ translate(
-					'Jetpack Search helps your visitors instantly find the right content – right when they need it.'
-				) }
-			/>
-		);
-	};
-
-	const renderSiteAntiSpam = () => {
-		return (
-			<JetpackBenefitsCard
-				headline={ translate( 'Anti-spam' ) }
-				description={ translate(
-					'Jetpack anti-spam automatically clears spam from comments and forms.'
-				) }
-			/>
-		);
-	};
-
-	const renderSiteActivity = () => {
-		return (
-			<JetpackBenefitsCard
-				headline={ translate( 'Activity Log' ) }
-				description={ translate(
-					'The Activity Log shows a list of management events that have occurred on your site.'
-				) }
-			/>
-		);
-	};
-
 	return (
 		<React.Fragment>
 			{
@@ -162,8 +67,22 @@ const JetpackBenefits: React.FC< Props > = ( { siteId, productSlug } ) => {
 					isStandalone={ isJetpackBackupSlug( productSlug ) }
 				/>
 			) }
-			{ productHasSearch( productSlug ) && renderSiteSearch() }
-			{ productHasAntiSpam( productSlug ) && renderSiteAntiSpam() }
+			{ productHasSearch( productSlug ) && (
+				<JetpackBenefitsCard
+					headline={ translate( 'Search' ) }
+					description={ translate(
+						'Jetpack Search helps your visitors instantly find the right content – right when they need it.'
+					) }
+				/>
+			) }
+			{ productHasAntiSpam( productSlug ) && (
+				<JetpackBenefitsCard
+					headline={ translate( 'Anti-spam' ) }
+					description={ translate(
+						'Jetpack anti-spam automatically clears spam from comments and forms.'
+					) }
+				/>
+			) }
 			{ siteHasScan() && productHasScan( productSlug ) && (
 				<JetpackBenefitsScanHistory
 					siteId={ siteId }
@@ -175,7 +94,14 @@ const JetpackBenefits: React.FC< Props > = ( { siteId, productSlug } ) => {
 				 * could look to expand output by using requestActivityLogs to get this information,
 				 * there is also an endpoint for /activity/counts that has no matching state components that could get set up
 				 */
-				productHasActivityLog( productSlug ) && renderSiteActivity()
+				productHasActivityLog( productSlug ) && (
+					<JetpackBenefitsCard
+						headline={ translate( 'Activity Log' ) }
+						description={ translate(
+							'The activity log shows a full list of management events that have occurred on your site.'
+						) }
+					/>
+				)
 			}
 		</React.Fragment>
 	);
